Align SemanticAnalysis to canonical types and print correct function return types (#26)
Changes
- Propagate function return type through semantic model (FunctionNode.ReturnType) and use it in both SemanticSnapshotPrinter and PrintNodeVisitor.
- Update SemanticAnalysis samples and verified snapshots to use canonical types: i16 variables and i16 main()/i16 other() headers.
- Refresh Verify snapshots (including double-prefixed files) to match canonical output.
- Fix Z80 tests to use new FunctionNode constructor signature.

Verification
- Full solution tests passing locally: total 76; passed 73; skipped 3; failed 0.

+++ src/RetroSharp.SemanticAnalysis.Tests/Samples/scopes/blocks/001_shadowing.c
@@ -1,1 +1,1 @@
-void main(){ int a; { int a; a = 1; } a = 2; }
+void main(){ i16 a; { i16 a; a = 1; } a = 2; }
